pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llmcost"
version = "0.1.0"
description = "CLI tool for comparing LLM API pricing, ranked by cost-effectiveness against LMSYS Arena scores"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["llm", "pricing", "api", "cost", "comparison", "openai", "anthropic", "deepseek"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]

dependencies = [
    "httpx>=0.27",
    "rich>=13",
    "pyyaml>=6",
    "beautifulsoup4>=4.12",
    "playwright>=1.40",
]

[project.urls]
Repository = "https://github.com/xiapuyang/llmcost"
Issues = "https://github.com/xiapuyang/llmcost/issues"

[project.optional-dependencies]
dev = ["pytest>=8", "respx>=0.21"]

[project.scripts]
llmcost = "llmcost.cli:main"

[tool.pytest.ini_options]
testpaths = ["llmcost/pricing/tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["llmcost*"]
